bugs

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/daos/UserDao.java
@@ -41,7 +41,7 @@ public interface UserDao {
 
     boolean updatePhoneOfUser(int userId, String phone);
 
-    boolean updateImageOfUser(int userId, int [] image);
+    boolean updateImageOfUser(int userId, byte [] image);
 
     boolean updateAddressOfUser(int userId, String address);
